Chore: move unit tests to standalone files for progress::*
